pyproject.toml
[project]
name = "mymasterproject"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bm3d>=4.0.3",
    "colour-demosaicing>=0.2.7",
    "exifread>=3.5.1",
    "h5py>=3.16.0",
    "ipykernel>=7.1.0",
    "ipympl>=0.9.8",
    "jupyterlab>=4.5.2",
    "matplotlib>=3.10.8",
    "numba>=0.63.1",
    "numpy>=2.3.5",
    "opencv-python>=4.11.0.86",
    "pandas-stubs~=2.3.3",
    "pyiqa>=0.1.14.1",
    "rawpy>=0.26.1",
    "scikit-image>=0.26.0",
    "scipy-stubs~=1.16.3",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "tqdm>=4.67.1",
]
